added a command line interface for sending appointment reminder emails
the function sends an email to appointments that occur within the next 24 hour peroid (if they have not yet been notified). to send reminders the database schema must first be updated:

php index.php cli update

after that, a cron job can be set up to run the following command:

php index.php cli send_appointment_reminders

+++ src/application/migrations/001_specific_calendar_sync.php
@@ -3,17 +3,26 @@
 class Migration_Specific_calendar_sync extends CI_Migration {
 
 	public function up() {
-		$fields = array(
-			'google_calendar' => array(
-				'type' => 'VARCHAR',
-				'constraint' => '128',
-				'null' => TRUE)
-		);
-		
-		$this->dbforge->add_column('ea_user_settings', $fields);
+
+		if (!$this->db->field_exists('google_calendar', 'ea_user_settings'))
+		{
+			$fields = array(
+				'google_calendar' => array(
+					'type' => 'VARCHAR',
+					'constraint' => '128',
+					'null' => TRUE)
+				);
+
+			$this->dbforge->add_column('ea_user_settings', $fields);
+		}
 	}
 
 	public function down() {
-		$this->dbforge->drop_column('ea_user_settings', 'google_calendar');
+		if ($this->db->field_exists('google_calendar', 'ea_user_settings'))
+		{
+			$this->dbforge->drop_column('ea_user_settings', 'google_calendar');
+		}
 	}
 }
+
+?>
